% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxMDS.R
\name{datasplit_surv}
\alias{datasplit_surv}
\title{Data Split for Survival Response}
\usage{
datasplit_surv(
  X,
  Y,
  MS,
  COV,
  penalty = c("MCP", "lasso", "SCAD"),
  q = 0.1,
  n_split = 25
)
}
\arguments{
\item{X}{a vector of the exposure.}

\item{Y}{a vector of the outcome.}

\item{MS}{a matrix of the mediators after the pre-liminary screening.}

\item{COV}{optional, a matrix of the potential covariates.}

\item{penalty}{the penalty used in the cox regression.}

\item{q}{the pre-defined FDR level in the data splitting procedure.}

\item{n_split}{the pre-defined number of multiple data splittings.}
}
\value{
A vector of mediator IDs selected by AKO.
}
\description{
Data Split for Survival Response
}
\section{Functions}{
\itemize{
\item \code{datasplit_surv()}: the data for survival outcome and compute the corresponding statistics.

}}
